fix(animation): address slice D1 review (clamp, version, complexity, ctrl-marquee)
CodeRabbit Major:
- moveKeyframes() now CLAMPS the requested batch dt to the largest
  legal magnitude instead of rejecting any out-of-bounds drag. A
  multi-selection sliding into a clip boundary lands on the nearest
  legal time. Refactored into clampBatchDt() + batchCollides()
  helpers so cognitive complexity drops well below Sonar's S3776
  threshold.
- selectInRect() takes an `additive` flag. timelineArea snapshots
  the modifier state at press time and passes it through to the
  marquee commit. Ctrl/Cmd+drag now extends the existing selection
  instead of replacing it. (CR Major)

Codex P2:
- pasteKeyframesAt() pre-checks collisions before pushing the
  PasteKeyframesCommand. If every entry would skip, no command is
  pushed onto the undo stack — keeps the history clean when the user
  accidentally pastes onto identical times.

CodeRabbit Minor:
- Reject unknown clipboard versions: paste payload's `version` field
  must equal 1, otherwise return 0.
- Reject incomplete entries: every TRS field must be present and
  numeric. Previously missing fields silently coerced to 0 / 1.
- BulkKeyframeCommands_test's hasKeyframeAt() now uses `<=` to match
  production's findKeyframeIndex tolerance check (was `<`, off by one).

SonarCloud:
- S5276 ×2: explicit static_cast<int> on QVector::size() in tr() count.
- S5276 ×1 in controller breadcrumb arg().
- (S5827 / S3776 in moveKeyframes resolved by the refactor above.)

Tests:
- MoveKeyframesShiftsAllByDt now also asserts source-times are gone +
  total keyframe count unchanged (catches faulty insert-without-remove
  implementations).
- MoveKeyframesRejectsClipBoundary renamed to
  MoveKeyframesClampsAtClipBoundary; adds a partial-clamp case that
  shifts a single keyframe to the nearest legal time.

+++ src/commands/BulkKeyframeCommands.cpp
@@ -70,7 +70,7 @@ MoveKeyframesCommand::MoveKeyframesCommand(Ogre::Skeleton* skeleton,
     , mItems(std::move(items))
     , mDt(dt)
 {
-    setText(QObject::tr("Move %n keyframe(s)", "", mItems.size()));
+    setText(QObject::tr("Move %n keyframe(s)", "", static_cast<int>(mItems.size())));
 }
 
 bool MoveKeyframesCommand::shiftAll(float fromOffset, float toOffset)
@@ -114,7 +114,7 @@ PasteKeyframesCommand::PasteKeyframesCommand(Ogre::Skeleton* skeleton,
     , mEntries(std::move(entries))
 {
     mApplied.resize(mEntries.size());
-    setText(QObject::tr("Paste %n keyframe(s)", "", mEntries.size()));
+    setText(QObject::tr("Paste %n keyframe(s)", "", static_cast<int>(mEntries.size())));
 }
 
 void PasteKeyframesCommand::redo()
